+++ lib/rubocop/cop/performance/detect.rb
@@ -3,24 +3,29 @@
 module RuboCop
   module Cop
     module Performance
-      # This cop is used to identify usages of `select.first` or
-      # `find_all.first` and change them to use `detect` instead.
+      # This cop is used to identify usages of
+      # `select.first`, `select.last`, `find_all.first`, and `find_all.last`
+      # and change them to use `detect` instead.
       #
       # @example
       #   # bad
       #   [].select { |item| true }.first
+      #   [].select { |item| true }.last
       #   [].find_all { |item| true }.first
+      #   [].find_all { |item| true }.last
       #
       #   # good
       #   [].detect { |item| true }
+      #   [].reverse.detect { |item| true }
       class Detect < Cop
-        MSG = 'Use `%s` instead of `%s.first`.'
+        MSG = 'Use `%s` instead of `%s.%s`.'
+        REVERSE_MSG = 'Use `reverse.%s` instead of `%s.%s`.'
 
         SELECT_METHODS = [:select, :find_all]
 
         def on_send(node)
           receiver, second_method = *node
-          return unless second_method == :first
+          return unless second_method == :first || second_method == :last
           return if receiver.nil?
 
           receiver, _args, _body = *receiver if receiver.block_type?
@@ -30,21 +35,28 @@ module RuboCop
 
           range = receiver.loc.selector.join(node.loc.selector)
 
-          add_offense(node, range, format(MSG,
+          message = second_method == :last ? REVERSE_MSG : MSG
+          add_offense(node, range, format(message,
                                           preferred_method,
-                                          first_method))
+                                          first_method,
+                                          second_method))
         end
 
         def autocorrect(node)
-          receiver, _first_method = *node
+          receiver, first_method = *node
 
+          replacement = if first_method == :last
+                          "reverse.#{preferred_method}"
+                        else
+                          preferred_method
+                        end
           first_range = node.loc.dot.join(node.loc.selector)
 
           receiver, _args, _body = *receiver if receiver.block_type?
 
           @corrections << lambda do |corrector|
             corrector.remove(first_range)
-            corrector.replace(receiver.loc.selector, preferred_method)
+            corrector.replace(receiver.loc.selector, replacement)
           end
         end