+++ spec/rubocop/cop/performance/detect_spec.rb
@@ -26,6 +26,13 @@ describe RuboCop::Cop::Performance::Detect do
         .to eq(["Use `detect` instead of `#{method}.first`."])
     end
 
+    it "registers an offense when last is called on #{method}" do
+      inspect_source(cop, "[1, 2, 3].#{method} { |i| i % 2 == 0 }.last")
+
+      expect(cop.messages)
+        .to eq(["Use `reverse.detect` instead of `#{method}.last`."])
+    end
+
     it "registers an offense when first is called on multiline #{method}" do
       inspect_source(cop, %(
         [1, 2, 3].#{method} do
@@ -37,6 +44,17 @@ describe RuboCop::Cop::Performance::Detect do
         .to eq(["Use `detect` instead of `#{method}.first`."])
     end
 
+    it "registers an offense when first is called on multiline #{method}" do
+      inspect_source(cop, %(
+        [1, 2, 3].#{method} do
+          |i| i % 2 == 0
+        end.last)
+      )
+
+      expect(cop.messages)
+        .to eq(["Use `reverse.detect` instead of `#{method}.last`."])
+    end
+
     it "registers an offense when first is called on #{method} short syntax" do
       inspect_source(cop, "[1, 2, 3].#{method}(&:even?).first")
 
@@ -44,7 +62,15 @@ describe RuboCop::Cop::Performance::Detect do
         .to eq(["Use `detect` instead of `#{method}.first`."])
     end
 
-    it "does not register an offense when #{method} is used without first" do
+    it "registers an offense when last is called on #{method} short syntax" do
+      inspect_source(cop, "[1, 2, 3].#{method}(&:even?).last")
+
+      expect(cop.messages)
+        .to eq(["Use `reverse.detect` instead of `#{method}.last`."])
+    end
+
+    it "does not register an offense when #{method} is used " \
+       'without first or last' do
       inspect_source(cop, "[1, 2, 3].#{method} { |i| i % 2 == 0 }")
 
       expect(cop.messages).to be_empty
@@ -72,12 +98,33 @@ describe RuboCop::Cop::Performance::Detect do
             )
           end
 
+          it "corrects #{method}.last to reverse.#{preferred_method} " \
+             '(with block)' do
+            new_source = autocorrect_source(
+              cop,
+              "[1, 2, 3].#{method} { |i| i % 2 == 0 }.last")
+
+            expect(new_source).to eq(
+              "[1, 2, 3].reverse.#{preferred_method} { |i| i % 2 == 0 }"
+            )
+          end
+
           it "corrects #{method}.first to #{preferred_method} (short syntax)" do
             new_source = autocorrect_source(
               cop,
               "[1, 2, 3].#{method}(&:even?).first")
 
             expect(new_source).to eq("[1, 2, 3].#{preferred_method}(&:even?)")
+          end
+
+          it "corrects #{method}.last to reverse.#{preferred_method} " \
+             '(short syntax)' do
+            new_source = autocorrect_source(
+              cop,
+              "[1, 2, 3].#{method}(&:even?).last")
+
+            expect(new_source)
+              .to eq("[1, 2, 3].reverse.#{preferred_method}(&:even?)")
           end
 
           it "corrects #{method}.first to #{preferred_method} (multiline)" do
@@ -90,6 +137,22 @@ describe RuboCop::Cop::Performance::Detect do
 
             expect(new_source).to eq(
               %([1, 2, 3].#{preferred_method} do
+                  |i| i % 2 == 0
+                end)
+            )
+          end
+
+          it "corrects #{method}.last to reverse.#{preferred_method} " \
+             '(multiline)' do
+            new_source = autocorrect_source(
+              cop,
+              %([1, 2, 3].#{method} do
+                  |i| i % 2 == 0
+                end.last)
+              )
+
+            expect(new_source).to eq(
+              %([1, 2, 3].reverse.#{preferred_method} do
                   |i| i % 2 == 0
                 end)
             )